Including private folder in the GLOBAL CMake configuration.
private folder is now included during the cmake configuration stage,
when a private code requires to be compiled we can call

make code_name

in the build directory. The executable will be created in the specified
directory in the CMakeList.txt that lives in the same folder as the
source code.

Check private/julio/dead_reckoning/CMakeList.txt file for an example