Applied low-pass filter for dead reacking Telit.
We applied a low-pass fiter to the dead reckoning test case

private/julio/dead_reckoning/demo_dead_reckoning_for_TelitSL869-DR.cpp

The "alpha" value was set to 0.90, this states to use 90% of the data of
the gyroscope and 10% of the date of the accelerometer when performing
data fusion. We can now the less-noisy signal and expect less
accumulated error.