Dead reckoning Telit: Get speed in knots from GPS.
We get the speed in knots from GPS and convert it to speed in m/s. Both
set of data are output to files.

Using the speed from GPS we compute the acceleration. This will be used
to compare with the body acceleration computed by the algorithm.

The speed from GPS is used to compute the position/trajectory. This will
be used to compare with the position computed by the algorithm.