Dead reckoning Telit: Working with acceleration lectures.
The idea is to have the lectures of the accelerometers as close as
possible to the values obtained from differentiation of velocity from
GPS. Currently the acceleration is shifted by a constant 0.2892.

Additionally, a kernel with more data is now used, we are using a 60
order kernel to convolve the input signals from the sensors.